+++ packages/axios-extension/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/axios-extension
 
+## 1.4.0
+
+### Minor Changes
+
+-   d2fd9a58: Enhanced LREP service to support merging of app descriptor variants
+
 ## 1.3.6
 
 ### Patch Changes